requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
Faker==0.7.4
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
PyHamcrest==2.1.0
pytest==8.3.4
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
structlog==25.1.0
typing_extensions==4.12.2
urllib3==2.3.0
